Refactored config to act like a utility + CARTLogic to minimize direct access.
